+++ loguru/__init__.py
@@ -11,7 +11,7 @@ from ._logger import Logger as _Logger
 
 __version__ = "0.2.5"
 
-logger = _Logger({}, None, False, False, False, False, 0)
+logger = _Logger({}, None, None, False, False, False, False, 0)
 
 if _defaults.LOGURU_AUTOINIT and _sys.stderr:
     logger.add(_sys.stderr)